program: output error for command line support
TEST: verify that an run error is output during execution of the
program.

Currently, a run error is thrown from the program unit run function.
For use with the command line compiler (ibc), the error needs to be
output in a format similar to compile errors in the compile source
function.  This could be added to the run function, however, for GUI
use, the error still needs to be thrown so that it can be caught by the
GUI and the error displayed via a pop up dialog.

Therefore, a new run code function as added that calls the run function,
and if a program error is thrown it is called and output (with a new
output error function) to that passed output stream.