program: handle a run error when at end of program
TEST: verify the correct error message when the bug stack error occurs
(at the end of the program with the automatic END command).

The error offset in this case is beyond the end of the program because
the automatic END command appended to the program is automatically
removed when the program stops running due to an error (or a normal
end).  This is no line associated with the error (the line string in
the program error class is empty), so a different error message is
require (one that doesn't contain the line number, column, etc.).

To make this test pass, a check for an empty line was added to the
output program error function.  If the line is empty, only the error
message is output.