program: cleaned up the error output functions
There were nearly identical output error functions, so these were
combined into a single function.  This was accomplished by adding a
program error class constructor that takes a reference to a compile
error instance and the appropriate values are extracted from it.  Also
added a type flag to the program error (compile or run error).  This is
used to return a type string used to differentiate between compile and
run error output.

Several variables and functions were changed from integer to unsigned
integer to be more compatible with the size type (not quite since it
is shorter, meaning that unsigned integer can't be initialized from a
size type).