program: changed command line compiler to report run errors
TEST: verify that the command line compiler will report a run error
correctly when one occurs.

Initially this test failed for two reasons, the ibc command was not
returning an error when a run error occurred and did not report the
error to the standard output.

These tests were made to pass by adding a success/fail status return
value to the program unit run code function and calling this from the
execute function instead of the plain run function and throwing an error
if the run code returns a fail (false) status.

For completeness, checks for the return value of the run code function
was added to the program unit tests.  To check for a success (true)
status, a new test was added with a successful running program.  This
test case was renamed since it now tests more than run errors.