program: separated out compile function with no output
Continuing the theme of having separate functions for the command line
compiler and future GUI, a new compile function was extracted from the
compile source function.  The new function adds any compile errors to a
vector of program errors (each being converted).  This function is meant
to be used by the GUI, which will take the program errors and display
them on the screen if the lines are visible.  This function does not
require an output stream argument.

The compile source function was modified to call the new compile
function.  If there are any program errors in the returned vector, they
are output to the passed output stream.  The returns success if the
program errors vector is empty.

The handle error function requires too much overhead (arguments and
return function) for little gain, so the code was moved to the new
compile function in the catch block.

Though not necessary to keep the tests passing, the unit tests were
modified to use the new compile function since on most uses, the output
stream was not used.