exponential: check for an overflow
TEST: verify that if the exponential causes the result that is larger
than will fix into an integer, that a run error is thrown.

With the multiple method for calculating the power on integers, there is
no built in overflow check when a multiplication overflows an integer,
the upper bits are simply lost.  For the calculation chosen for this
test, 2 ^ 31, the result is actually a negative number, which is wrong.

To make this test pass, the result variable in the calculate power was
changed to a 64-bit variable.  After the multiplication is performed,
the result is checked to see if it is still within the limits of an
integer, and if not, an overflow run error is thrown.