exponential: check for an underflow with a negative number
TEST: verify that if the exponential operator causes the result that is
a negative number larger than will fit into an integer, that a run error
is thrown.

This test initially failed because the code was only checking if the
result was greater than the large integer value.  To make this test
pass, the result needs to be compared to lowest (minimum) integer value.

For efficiency, a separate function was created for handling the
minimum checks, and the other function handles maximum checks.  The
calculate power function checks the value and calls the apprpriate
function.