exponential: check for an overflow with a negative number
TEST: verify that if the exponential operator with a negative number
causes the result a positive number larger than will fit into an
integer, that a run error is thrown.

Negative numbers that are raised to an even power will have a positive
result, so positive overflows also need to be checked for.

This test initially failed because the code was only checking if the
result was less than the smallest integer value.  To make this test
pass, the result also needs to be compared to highest (maximum) integer
value.

This was done as a separate test, with the other part of the test
implemented separated so that both parts of the check have a test.