exponential: use alternate power function for high exponents
TEST: verify the new alternate is executed when the exponents is greater
than 18.

Once the exponents give above a certain value (found to be above around
18) that the multiply method ends up consuming more time then converting
the numbers to doubles, use the standard power function and converting
back to an integer.

So only if the exponent is less than 19 will the multiple method be used
for positive values, otherwise the standard power function will be used.
This will also prevent a really high exponent from being used with a
zero or one that will not overflow and will take a very long time to
multiply the result.  In the end though, this method only helps with
values of three or less since even with three, an exponent of 20 or
higher will overflow (for higher numbers, the lower exponent will
overflow).

To initially verify failure, coverage was checked and the new function
was being executed before a new test was added because the previous test
for 2^31 was routed to the new function.  This test was changed to 16^8
to execute the older code that was suppose to be tested (an overflow
check).