exponential: use alternate power for high exponents with negative values
TEST: verify the alternate is executed when the exponent is greater than
16 with a negative value.

Once the exponents give above a certain value (found to be above around
16 with negative values - lower for positive since there is an extra
overflow check) that the multiply method ends up consuming more time
then converting the numbers to doubles, use the standard power function
and converting back to an integer.

So only if the exponent is less than 17 will the multiple method be used
for negative values, otherwise the standard power function will be used.
This will also prevent a really high exponent from being used with a
negative one that will not overflow and could take a very long time to
multiply the result.  In the end though, this method only helps with
negative values up to three and even with three, an exponent of 20 or
higher will overflow (for higher numbers, the lower exponent will
overflow).

To initially verify failure, coverage was checked and the new function
was not being executed before a new test was added.