exponential: underflow check with alternate power with negative values
TEST: verify that if the exponential operator with a negative number
causes the result a positive number less than will fit into an integer,
that a run error is thrown.

Negative numbers that are raised to an odd power will have a negative
result, so negative overflows (underflow) also needs to be checked for.

This test initially failed because the code was only checking if the
result was greater than the largest integer value.  To make this test
pass, the result also needs to be compared to lowest (minimum) integer
value.

This was done as a separate test, with the other part of the test
implemented separated so that both parts of the check have a test.