exponential: cleaned up power integer-integer code
The power integer-integer grew some what large, so it was moved to a
new header file.  All of the code was put into the header file so the
entire function can be in lined where called.

The repeated overflow checks were extracted into separate functions.
These functions needs to be template functions since the result
variables are both a 64-bit integer and a double.