exponential: implemented double-double operator execute function
TEST: verify that the exponential double-double operator gives the
correct result.

To make this test pass, the execute exponential double-double function
was implemented, which pops the exponent value from the stack and calls
the standard power function with the value on top of the stack putting
result on top of the stack.  A pointer to this function replaces the
null pointer in the exponential double-double code instance.