exponential: check for a domain error with double-double operator
TEST: verify that a domain error is thrown when a negative number is
raised to a non-integer exponent.

Under the hood, the standard power function probably takes the logarithm
of the value, multiplies it be the exponent and then takes the inverse
logarithm.  Logarithms cannot be taken of negative values (it probably
uses an alternate calculation with an integer exponent).

To make this test pass, the result of the power function is placed into
a temporary result variable, which is checked to see if is a NAN
(not-a-number), and throws an error if it is.