exponential: check for an overflow error with double-double operator
TEST: verify that an overflow error is thrown when the result is too
large to fit into a double value.

The power function returns a special huge value if an overflow has
occurred.  The result is checked for the value for which an overflow
error is thrown.

The extra checks in the unit tests for if a program error is thrown was
unnecessary since the run code function catches errors and produces the
appropriate error message, therefore these extra checks were removed.