exponential: divide by zero error check with double-double operator
TEST: verify that a divide by zero error is thrown when a value of zero
is raised by a negative value.

The power function also returns the special huge value if a value of
zero is raised by a negative (in other words, infinity).  To distinguish
this error with an overflow error, the value is checked for a zero
value.

Since these three checks were starting to clutter the execute
exponential double-double function, the checks were moved to a new
validate power result function.