exponential: implemented integer-double operator execute function
TEST: verify that the exponential integer-double operator gives the
correct result.

To make this test pass, the execute exponential integer-double function
was implemented, which pops the exponent value from the stack, and gets
the integer value from the top of the stack.

Since the rest is nearly identical to the double-double operator (except
converting the integer value to double), this code was extracted to a
new calculate power double-double function, which contains a double
value argument (and this takes care of the conversion from integer to
double).  A pointer to the new execute function replaces the null
pointer in the exponential double-double code instance.

Since this function is using the same code and therefore the same error
checks, no additional tests are needed for the integer-double
exponential operator.