exponential: implemented double-integer operator execute function
TEST: verify that the exponential double-integer operator gives the
correct result.

To make this test pass, the execute exponential double-integer function
was implemented, which pops the integer exponent value from the stack,
and gets the double value from the top of the stack.  A pointer to the
new execute function replaces the null pointer in the exponential
double-integer code instance.

Since taking the exponential of a double to an integer is going to be
similar to an integer to an integer, a new power double-integer function
operator class was created.  For now this class just has a constructor
and function operator function that uses the multiply method for
calculating the power.  For efficiency, all functions will in line
functions defined in the header file.

OTHER: some of the wording in the other exponential tests was wrong and
was corrected.