exponential: overflow check for double-integer operator
TEST: verify that an overflow error is thrown if the result it too large
to fit into a double.

To make this test pass, a check for the special huge value was added to
the multiple for positive exponent function at doing the
multiplications.  It is not necessary to check after each multiply
because once a double value goes to the huge value, it stays set to the
huge value.  This could result in extra multiplications, but the number
of multiplications won't be that high since the standard power function
will be used for larger exponents like was done for the integer-integer
operator.