exponential: use alternate power for high exponents
TEST: verify the alternate power is executed when the exponent is not
less than 19.

Once the exponents give above a certain value (found to be above around
18), the multiply method ends up consuming more time than converting the
exponent from integer to double and using the standard power function.

So only if the exponent is less than 19 will the multiple method be
used, otherwise the standard power function will be used.

To initially verify failure, coverage was checked and the new function
was not being executed before a new test was added.