exponential: overflow check with alternate power
TEST: verify that an overflow error is thrown when the alternate power
function for double-integer.

To make this test pass, the result after calling the standard power
function, a check was added if the value is the huge value, an overflow
error is thrown.

OTHER: removed the checks for false on the return value from the run
code function calls.  These are not really necessary as the check for
the output stream string is sufficient to check if there was an error or
not.