exponential: use alternate power for high negative exponents
TEST: verify the alternate power is executed when the exponent is not
greater than -17.

Once the exponents go above a certain value (found to be above around
-16), the divide method ends up consuming more time than converting the
exponent from integer to double and using the standard power function.

So if the exponent is greater than -17, the divide method will be used,
otherwise the standard power function will be used (which is in a new
function).

OTHER: The use of isnan() function should have been put in the standard
name space (which in fact is necessary on Windows).  The value used in
the new unit test was chosen to produce a three digit exponent in the
result so that it also works on Windows (which always produces three
digit exponent, where as Linux only produces three if needed otherwise
there are only two).