exponential: overflow check with alternate power and negative exponents
TEST: verify that an overflow error is thrown when the alternate power
function for double-integer with negative exponents.

Unlike with the integer-integer exponential, an underflow can't happen
since the value just goes to zero.  However, an overflow can occur if
the number being raised is really small and the exponent is negative.

To make this test pass, the result after calling the standard power
function, a check was added if the value is the huge value, an overflow
error is thrown.

Since the two use power functions ended up the same, one was removed and
the other renamed appropriately.  Since there are checks against the
special huge value, this check was put into another function.