parens: parse an opening parenthesis character
TEST: verify that an opening parenthesis character is not rejected as an
invalid character.

A parenthetical expression consists of an opening parenthesis, a numeric
expression, and a closing parenthesis; and can be anywhere that an
operand can be found in an expression.

To make this test pass, the contents of the compile numeric operand
function was moved to a new compile numeric constant function except for
the statement that gets the current column.  The new function then first
peeks at the next character to see if it is an opening parenthesis
character.  Otherwise it calls the compile numeric constant, the same
functionality as before.

If an opening parenthesis character is found, the character is consumed,
but then just returns a double data type as if an expression was
compiled, which was all that was necessary to make this first test pass.

The column needs to be obtained in this new function since the next
character is peeked for an opening parenthesis, which will invalidate
the column of the compiler if there is no character to peek (this
problem was found when the "-" negation test started to fail because the
column was not currect).