parens: compile expression inside the parentheses
TEST: verify that the expression inside the parentheses is compiled.

This test initially failed because the expression inside the parentheses
was not being compiled (only the opening parenthesis was consumed).  The
test only consists of verify that the expected code length to account
for all the codes that should be in the code.  The actually contents
will be verified later when the expression is executed.

To make this test pass, after consume the opening parenthesis, the
compile numeric operand function recursively calls the compile numeric
expression function, and returns the data type returned.

What is not technically correct is that the double data type is passed
as the expected data type to the compile numeric expression function.
This is sufficient for making the test compile, but this data type will
need to be changed to a generic numeric data type (which doesn't exist
yet) because eventually the compile numeric expression will need to add
a conversion code if the data type doesn't match (integer or double) the
data type expected (requested).