parens: parse the closing parenthesis character
TEST: verify that the closing parenthesis character is consumed at the
end of a parenthetical expression.

The test was checking if the end-of-file character is the next character
to be processed.  This test initially failed because the character at
the end was not being consume.

To make this test pass, the next character is consumed (checking for the
actual closing parenthesis was not yet necessary).

OTHER: corrected the indenting in the unit test (somehow QtCreator was
doing something strange with the auto-indenting).