project: made use of auto were possible and some other clean ups
Replaced types on local variable declaration with the use of C++11
'auto' that automatically determines the type.  This required the use
of some static casts to force auto to determine the correct type with
a conversion.  This also makes it clear, and not implicit, that a
conversion is occurring.  Added descriptive named constants to some
literal constants to explain what the purpose of the numbers are.
Extracted the function for compiling parentheses expressions.  And fixed
the unit test for skipping white space after an opening parenthesis (the
last test added), which should have been checking for an end-of-file
(to make sure the entire expression was compiled).