parens: check for a closing parenthesis
TEST: verify an error is thrown in there is no closing parenthesis
character after the numerical expression following an open parenthesis.

This test initially failed because the compile parentheses function was
simply getting the next character after the numerical expression
assuming it was a closing parenthesis.

To make this test pass, a check was added if the next character is not
a closing parenthesis, then throw a compile error (for now just a blank
error).