parens: added message and column to missing closing parenthesis error
TEST: verify the correct message, column and length of the missing
closing parenthesis error.

To make this test pass, a message was added to the compile error thrown.
The current column was also added to the compile error.  The column
needed to be obtained before peeking at the next character in the is
operator character function, which is where the closing parenthesis
character will first be seen.