compiler: cached peek character to protect column
A pattern was seen where the expression compiler needed to get the
current column more often from the compiler before peeking at the next
character.  This was because the column (the current get position in the
stream) becomes invalid once the stream is peeked at the end of the
stream.  This would only continue as the expression compiler is further
implemented.

Rather that continue this pattern, which will surely become quite
complex, the compiler class was modified where a new peek character
member (the cache) was added to hold the peek next character.

The peek next character function was modified to only get the current
column (stream get position) and peek at the next character, if the peek
character member is not set (a zero is considered not set).  Otherwise
the cached peek character is returned.

The get next character function was modified to reset the peek character
cache (since the peek character is now used).

The skip white space function was modified to only skip white space if
the peek character is not the end of file character, and the peek
character is also reset (the cache can't be reset if at the end of the
stream, otherwise the peek character function will get the invalid
stream get position).

The expression compiler no longer needs to maintain a copy of the
current column, so this member was removed.  These changes should
simplify the expression compiler going forward.