recreator: created unary and binary operator functions
The functionality for recreating unary and binary operators was moved
from the recreate negate and exponential functions and was made generic
where the keyword of the operator (symbols for these two operators) is
passed in as an argument.  The recreate negate and exponential functions
were modified to call these functions passing in the appropriate keyword
string.

This seemed to be an appropriate refactoring since most of the
statements was calling recreator member functions.  This changes was
made because the operators will be doing the same thing except that the
operator keyword will be different for each.  These changes are in
preparation for the recreation of parentheses.