precedence: implemented a class to define operator precedences
For the IBCP, precedences were assigned by number in the massive table
of data.  Care was required to make sure the precedences were right
between the difference operators and that each of the codes for the
similar operators (with different data types) were the same.  There is a
high chance for mistakes using this scheme.

For the IBCX, a better scheme was developed where precedences will be
assigned automatically including to each of the different operator data
types.  For compilation, the precedences are currently implied by the
order the functions for each operator group are called (the so caller
recursive descent method).

Parentheses override the normal precedences, but nothing is stored to
indicate where the parentheses were in the expression.  Unlike with the
IBCP, unnecessary parentheses will be removed (not recreated).  This is
unnecessary and requires extra work to accomplish.  Parentheses will
only be recreated as needed by the precedences of the operators.

Therefore, to recreate parentheses, operators need to be assigned
precedences.  When a higher precedence operator is recreated with its
operands, if either operand is an operator with a lower precedence then
parentheses need to be added around the operand.

A new precedence class was implemented for maintain the precedences of
the operators.  This class will give access to the precedences of the
operators (by code value), and will give access to the keywords for the
operators, which will be stored with the precedences.

Inside the source file for this precedence class, there is a private
precedence info class that will actually hold maps from code value to
precedence and keyword.  Since precedences are not needed just yet, the
instance of this [singleton] class just holds the operator keywords,
which contains an unordered map of code value to keyword.  This map is
initialized in the constructor for this class, which is defined as
private.  The single instance of this class is contained in the static
get instance function (while will create and initialize the instance the
first time it is called.

This is private operator data structure defined in this class containing
a reference to the operator codes and keyword for an operator.  The
constructor contains an initializer list of these operator data
structures, which is used to initialize the maps (just keywords for
now).

To support a single kind of codes reference in the operator data
structure, a new operator codes base class was added, which contains a
single pure virtual function for getting a vector of code values for the
codes contained within.  Both the unary and numeric operator codes
classes are now derived from this class and the code values override
functions were implemented that creates and returns a vector of code
values.

A code value member was added to the recreator class, which is set by
the recreate one code function after reading a code from the program.  A
get operator keyword was also added to use this code value to get the
keyword for the operator (using the new precedence class).

The recreate negate and and exponential functions were modified to get
the keyword from the recreator (via the precedence class).  This is a
further step in making the operator recreate functionality generic.