recreator: made unary and binary operator functions fully generic
The recreate negate and exponential functions were renamed to recreate
unary and binary, and were moved to the recreator source file.  The
getting of the operator keyword was moved into the recreator member
functions of the same name and the keyword arguments removed.  These
standalone recreate functions call the recreator functions and are meant
to be used in the code instances where the recreate member functions
cannot be used.