recreator: add parentheses to left side operand if needed
TEST: verify that an expression with parentheses around the left side
operand are recreated.

To make this test pass, a number of changes were needed including adding
precedences to the precedence class, a precedence member with the string
member in the recreate stack item with needed access functions, and
modifying the recreate binary function to compare the precedences of the
operator being recreated and that of the left side operand.

A get precedence function was added to the precedence class.  A code
value to precedence map was added to the precedence info class along
with an access function to get a precedence for a code value.  A current
precedence was also added, which is initialized to 1 (skipping over
zero, which is the highest precedence that is for non-operator operands
like constants), and incremented for each operator data entry that is
processed during initialization.  Each code value or an operator is then
mapped to this precedence.

The top function of the recreator was renamed to top string and a new
top precedence function was added.  A set top precedence function was
added to set the precedence of the item on top of the stack.  The push
function used to push operands on the stack was modified to assign a
precedence of 0 (highest) when adding the string to the stack.  A get
precedence function was added to call the get precedence function of the
precedence class.

The recreator recreate binary operator function was modified to check
if the left precedence if higher than the operator precedence to add
parentheses around the left operand.