negate: corrected precedence of exponential to negation
TEST: verify that the exponential operator is higher precedence than the
negation operator.

The test expression "2.0^ - 2^3" should give the answer 0.00390625 since
the "2^3" should be evaluated first, then negated (-8), which then is
the exponent to raise 2.0 to.  This test initially failed giving the
incorrect answer 0.015625 (2.0 raised to -2 raised to 3).

This problem was discovered during testing the recreation precedences of
these operators (which was not working correctly).  This problem was
corrected by changing the expression compiler compile negation function
to call the compile exponential function instead of the compile numeric
operator function.  This will give the exponential operator a higher
precedence than the negation operator (exponent operator binds first).