recreator: add parentheses to right side operand if needed
TEST: verify that an expression with parentheses around the right side
operand are recreated.

To make this test pass, the recreate binary operator function was
modified where the right side precedence is obtained when the right side
operand string is obtained.  When appending the right side operand,
parentheses are added around it if the right side precedence is higher
or the same (because operators normally group left to right) as the
current operator precedence.  The top precedence is set to the operator
precedence.