recreator: do not add parentheses on right side operand unnecessarily
TEST: verify when a negate operator is in the middle of two exponential
operators, that the recreator does not add parentheses around the second
exponential operator expression.

This issue was discovered in the process of correcting the problem with
the compile negation function.  With the code now correctly compiled for
the expression "3^- 1^2", the recreator was incorrectly adding
parentheses around negate and second exponential operators.

To correct this issue (solution found in the IBCP code), a unary
operator flag was added to the recreator stack item.  This flag is
normally set to false, but is set to true when a unary operator
expression is put on top of the stack.  In the recreate binary operator
function, this flag is obtained for the right side operand.  Then when
checking for lower precedence, this flag was not be set (in other words,
the presence of a unary operator will prevent parentheses from being
added around the operator.