parens: add parentheses around left side if there is a unary operator
TEST: verify that parentheses are added around the left side if there is
a unary operator on the left side.

The presence of a unary operator on the left side interferes with the
normal checking of the precedence of the higher precedence exponential
operator.  Previously parentheses would not be added since both
exponential operators are the same precedence.  For an expression like
"3^(- 1)^2" or "(3^- 1)^2" (both equivalent), the parentheses need to be
added or the second exponential operator will bind the operands over the
negate operator.

This test initially failed (no parentheses were added).  To get this
test to pass, the recreate binary [exponential] operator needs to pass
the unary flag on, which then needs to be checked for when the second
binary operator compares to the left side precedence.

At the end of the recreate binary operator function, if the right side
operand has the unary flag set, then the top unary operator flag [of the
binary operator] is also set.  Then when the left side precedence is
checked for lower precedence to add parentheses, parentheses will also
be added if the left side unary operator flag is set.

Curiously, the first expression is recreated as the second (the
parentheses are moved) - both are equivalent.  This action will be left
as is as the recreated expression more clearly shows the order of the
operations.  Also worthy of now, the IBCP did not handle this situation
correctly and completely removed the parentheses on the first expression
(though the second expression was handled correctly).