parens: prevent unnecessary parentheses when unary operator is present
TEST: verify when a unary operator is present, that extra parentheses
are not recreated.

The test expression "(- 3)^1^2" initially recreated extra parentheses
around the first exponential operator due to a bug.  The bug was that
when the left side operand contained the unary operator flag, this flag
was not being reset at the end of the recreate binary operator function,
and with the next call, the unary operator flag was still set causing
the extra set of parentheses to be recreated.

To correct this bug, the set top unary operator access function was
given a new value argument for the unary operator flag (so that the flag
can also be cleared).  The recreate unary operator function was modified
to pass a true value.  The recreate binary operator function was
modified to just call the set top unary operator function with the value
of the right unary operator flag variable.

There are more changes that are needed with parentheses and checking
precedences, but these changes cannot be done until another operator is
implemented that has a lower precedence than the exponential and negate
operators.  This concludes parentheses implementation for now.