multiply: started implementation of product expressions
TEST: verify that a simple multiplication expression is parsed to the
end of the expression.

This test was put into a new multiply unit test source file.  To make
this test pass, a new compile product function was added to the
expression compiler.  It was named 'product' since it will also handle
the divide operator, which is at the same precedence level as multiply.

The compile numeric expression was changed from calling the compile
exponential function to calling the new compile product function.  All
that was needed to make the test pass (and all the other existing unit
tests) is for the compile product function to call the compile
exponential function for the left side operand, checking if the next
operator character is an asterisk (the is operator character function
takes care of skipping white space around the operator).  If the
operator is multiply, then the compile exponential function is called
again for the right side operand.  No other action is needed for now.
The left side data type is returned.