multiply: implemented compilation of the rest of the multiply codes
TESTS: verify that with two double, integer/double, and double/integer
operands, that the appropriate multiply operator codes are added after
the operand codes.

To speed up the implementation of the rest of the multiply codes, and
since the implementation is the same as the exponential operators, the
rest of the multiply code instances was added along with the operator
codes instance.  The compile product function was modified to work
similar to the compile exponential function except the if statement
remains (instead of the while) and the check for a null data type from
the right side operand, which won't occur since the compile exponential
throws an error for this condition.