multiply: implemented recreation of the rest of the multiply codes
TESTS: verify that a multiply expression with two double,
integer/double, and double/integer operands are recreated correctly.

Again to speed up the implementation of the rest of the multiply codes,
tests were added for the remaining multiply codes one at a time and the
recreate function pointer of the associated code instance was changed
from a null pointer to the recreate binary operator function.