multiply: don't add parentheses with higher precedence negate operator
TEST: verify when there is a negate operator on one of the operators of
a multiply operator, that parentheses are not added around the multiply
expression.

This test was initially failing because of an earlier change to get
parentheses added around an exponential.  This change carried the unary
operator flag to the binary (exponential) expression on the stop of the
recreator stack.

To make this test pass, the unary operator flag is only carried to the
binary operator if the unary operator precedence is lower than the
binary operator precedence (which for the exponential operator, it is,
not is not with the multiply operator).