multiply: recreate parentheses with higher precedence unary operator
TEST: verify parentheses are recreated around a multiply expression with
a higher precedence unary operator.

This test initially failed because the recreate binary operator did not
check precedence, which wasn't necessary with the higher precedence
exponential operator.

To make this test pass, the recreate unary operator was modified to
check the precedence of the operand against the precedence of the unary
operator, and if lower, then parentheses are added around the operand as
it is appended to the top of the recreator stack.  If lower precedence,
then the check for adding a space if the first character of the operand
is a digit or period is not necessary as the opening parentheses will
separate the operator from the constant.