multiply: implement integer-integer operator execute function
TEST: verify that the multiply integer-integer operator gives the
correct result.

To make this test pass, the multiply integer-integer function was
implemented, which pops the right hand side value from the stack,
multiplies it by the value on top of the stack and places the result
in the integer value on top of the stack.  The pointer to this function
replaced the null pointer of the integer-integer multiply code instance.